part1/src/requirements.txt
Django>=3.0
psycopg2-binary>=2.7,<3.0
djangorestframework==3.10.1
drf-yasg==1.17.0
django_money==1.0
Pillow==6.2.1
awscli==1.16.309
numpy==1.18.0
